add tests and docs

+++ lib/atech/object_store/file.rb
@@ -8,6 +8,9 @@ module Atech
       ## Raised if a file cannot be added
       class ValidationError < Error; end
       
+      ## Raised if a frozen file is editted
+      class CannotEditFrozenFile < Error; end
+      
       ## Returns a new file object for the given ID. If no file is found a FileNotFound exception will be raised
       ## otherwise the File object will be returned.
       def self.find_by_id(id)
@@ -30,7 +33,7 @@ module Atech
       
       ## Inserts a new File into the database. Returns a new object if successfully inserted or raises an error.
       ## Filename & data must be provided, options options will be added automatically unless specified.
-      def self.add_file(filename, data, options = {})
+      def self.add_file(filename, data = '', options = {})
         ## Create a hash of properties to be for this class
         options[:name]          = filename
         options[:size]        ||= data.bytesize
@@ -40,11 +43,10 @@ module Atech
         
         ## Ensure that new files have a filename & data
         raise ValidationError, "A 'name' must be provided to add a new file" if options[:name].nil?
-        raise ValidationError, "'data' must be provided to add a new file" if options[:blob].nil?
         
         ## Encode timestamps
-        options[:created_at] = options[:created_at].utc.strftime('%Y-%m-%d %H:%M:%S')
-        options[:updated_at] = options[:updated_at].utc.strftime('%Y-%m-%d %H:%M:%S')
+        options[:created_at] = options[:created_at].utc
+        options[:updated_at] = options[:updated_at].utc
         
         ##Create an insert query
         columns = options.keys.join('`,`')
@@ -55,9 +57,10 @@ module Atech
         self.new(options.merge({:id => ObjectStore.backend.last_id}))
       end
       
-      ## Initialises a new File object with the hash of attributes from a MySQL query
+      ## Initialises a new File object with the hash of attributes from a MySQL query ensuring that
+      ## all symbols are strings
       def initialize(attributes)
-        @attributes = attributes
+        @attributes = parsed_attributes(attributes)
       end
       
       ## Returns details about the file
@@ -95,6 +98,11 @@ module Atech
         @attributes['blob']
       end
       
+      ## Returns whether this objec tis frozen or not
+      def frozen?
+        !!@frozen
+      end
+      
       ## Downloads the current file to a path on your local server. If a file already exists at
       ## the path entered, it will be overriden.
       def copy(path)
@@ -103,33 +111,58 @@ module Atech
       
       ## Appends data to the end of the current blob and updates the size and update time as appropriate.
       def append(data)
+        raise CannotEditFrozenFile, "This file has been frozen and cannot be appended to" if frozen?
         ObjectStore.backend.query("UPDATE files SET `blob` = CONCAT(`blob`, #{self.class.escape_and_quote(data)}), `size` = `size` + #{data.bytesize}, `updated_at` = '#{self.class.time_now}' WHERE id = #{@attributes['id']}")
+        reload(true)
       end
       
       ## Overwrites any data which is stored in the file
       def overwrite(data)
+        raise CannotEditFrozenFile, "This file has been frozen and cannot be overwriten" if frozen?
         ObjectStore.backend.query("UPDATE files SET `blob` = #{self.class.escape_and_quote(data)}, `size` = #{data.bytesize}, `updated_at` = '#{self.class.time_now}' WHERE id = #{@attributes['id']}")
+        @attributes['blob'] = data
+        reload
       end
       
       ## Changes the name for a file
       def rename(name)
+        raise CannotEditFrozenFile, "This file has been frozen and cannot be renamed" if frozen?
         ObjectStore.backend.query("UPDATE files SET `name` = #{self.class.escape_and_quote(name)}, `updated_at` = '#{self.class.time_now}' WHERE id = #{@attributes['id']}")
+        reload
       end
       
       ## Removes the file from the database
       def delete
+        raise CannotEditFrozenFile, "This file has been frozen and cannot be deleted" if frozen?
         ObjectStore.backend.query("DELETE FROM files WHERE id = #{@attributes['id']}")
+        @frozen = true
+      end
+      
+      ## Reload properties from the database. Optionally, pass true to include the blob
+      ## in the update
+      def reload(include_blob = false)
+        query = ObjectStore.backend.query("SELECT #{include_blob ? '*' : '`id`, `name`, `size`, `created_at`, `updated_at`'} FROM files WHERE id = #{@attributes['id']}").first
+        @attributes.merge!(parsed_attributes(query))
       end
             
       private
       
+      def parsed_attributes(attributes)
+        attributes.inject(Hash.new) do |hash,(key,value)|
+          hash[key.to_s] = value
+          hash
+        end
+      end
+      
       def self.escape_and_quote(string)
+        string = string.strftime('%Y-%m-%d %H:%M:%S') if string.is_a?(Time)
         "'#{ObjectStore.backend.escape(string)}'"
       end
       
       def self.time_now
         Time.now.utc.strftime('%Y-%m-%d %H:%M:%S')
       end
+      
       
     end
   end
